uint64 => uint32

+++ test/test_dfs_tree.py
@@ -19,11 +19,11 @@ def test_c_update_nodes():
   hist_sum_sqs = np.random.random(size=(max_nodes, cols, vals)).astype(np.float64)
 
   py_node_gains = np.full(max_nodes, -np.inf, dtype=np.float64)
-  py_split_cols = np.zeros(max_nodes, dtype=np.uint64)
+  py_split_cols = np.zeros(max_nodes, dtype=np.uint32)
   py_split_bins = np.zeros(max_nodes, dtype=np.uint8)
 
   c_node_gains = np.full(max_nodes, -np.inf, dtype=np.float64)
-  c_split_cols = np.zeros(max_nodes, dtype=np.uint64)
+  c_split_cols = np.zeros(max_nodes, dtype=np.uint32)
   c_split_bins = np.zeros(max_nodes, dtype=np.uint8)
 
   py_update_node_splits(hist_counts, hist_sums, hist_sum_sqs, py_node_gains, py_split_cols, py_split_bins, 0)
@@ -110,7 +110,7 @@ def test_eval_tree():
   #
   tree = Tree(
     node_count = 5,
-    split_cols = np.array([1,   0,  0, 0, 0], dtype=np.uint64),
+    split_cols = np.array([1,   0,  0, 0, 0], dtype=np.uint32),
     split_vals = np.array([100, 0, 10, 0, 0], dtype=np.float32),
     left_children  = np.array([1, 0, 3, 0, 0], dtype=np.uint16),
     node_means     = np.array([0, 1, 0, 3, 2], dtype=np.double),
